rename primitive objects
renamed:    src/typer/cp-object/Float64.ts -> src/typer/cp-object/Float.ts
renamed:    src/typer/cp-object/Int16.ts -> src/typer/cp-object/Integer.ts
renamed:    src/typer/cp-object/SolidBoolean.ts -> src/typer/cp-object/Boolean.ts
renamed:    src/typer/cp-object/SolidNull.ts -> src/typer/cp-object/Null.ts
renamed:    src/typer/cp-object/SolidNumber.ts -> src/typer/cp-object/Number.ts
renamed:    src/typer/cp-object/SolidString.ts -> src/typer/cp-object/String.ts

+++ src/typer/cp-type/TypeInteger.ts
@@ -12,12 +12,12 @@ export class TypeInteger extends Type {
 	override readonly isBottomType: boolean = false;
 	override readonly isTopType:    boolean = false;
 	private constructor () {
-		super(false, new Set([OBJ.Int16.ZERO]));
+		super(false, new Set([OBJ.Integer.ZERO]));
 	}
 	override toString(): string {
 		return 'int';
 	}
 	override includes(v: OBJ.Object): boolean {
-		return v instanceof OBJ.Int16;
+		return v instanceof OBJ.Integer;
 	}
 }
