+++ src/validator/astnode-cp/utils-private.ts
@@ -39,7 +39,7 @@ export function oneFloats(t0: TYPE.Type, t1: TYPE.Type): boolean {
 
 
 
-export function valueOfTokenNumber(source: string, config: CPConfig): OBJ.Int16 | OBJ.Float64 {
+export function valueOfTokenNumber(source: string, config: CPConfig): OBJ.Integer | OBJ.Float {
 	const [cooked, is_float]: [number, boolean] = Validator.cookTokenNumber(source, config);
-	return (is_float) ? new OBJ.Float64(cooked) : new OBJ.Int16(BigInt(cooked));
+	return (is_float) ? new OBJ.Float(cooked) : new OBJ.Integer(BigInt(cooked));
 }
